+++ nd2k/types.py
@@ -2,9 +2,9 @@ from collections import defaultdict
 from dataclasses import dataclass
 from datetime import datetime
 from enum import Enum
-from typing import NamedTuple
 from .operation import OperationType, Operation
 from .nontrade import NonTrade
+from .trade import Trade
 
 
 CSV = list[list[str]]
@@ -23,33 +23,6 @@ class KoinlyTag(Enum):
 	SWAP            = "swap"
 	EXCHANGE        = "exchange"
 	EXCHANGE_FEE    = "fee"
-
-
-class TradingPair(NamedTuple):
-	"""
-	In a trading pair like BTC/EUR:
-	- Base asset (BTC): The asset being bought or sold.
-	- Quote asset (EUR): The asset used to price the base asset.
-	"""
-	base:  str
-	quote: str
-
-
-@dataclass
-class Trade:
-	summary:      str
-	trading_pair: TradingPair
-	base_asset:   Operation
-	quote_asset:  Operation
-	trading_fee:  Operation
-
-	@property
-	def date(self) -> datetime:
-		return self.base_asset.date
-
-	@property
-	def type(self) -> OperationType:
-		return self.base_asset.type
 
 
 @dataclass
@@ -109,26 +82,3 @@ class PartialExchange:
 
 Transaction = Trade | NonTrade | Swap | Exchange
 TransactionGroups = defaultdict[str, list[Transaction]]
-
-
-@dataclass
-class PartialTrade:
-	summary:      str
-	trading_pair: TradingPair
-	base_asset:   Operation | None = None
-	quote_asset:  Operation | None = None
-	trading_fee:  Operation | None = None
-
-	@property
-	def type(self) -> OperationType:
-		return self._any_asset.type
-
-	@property
-	def _any_asset(self) -> Operation:
-		any_asset = self.base_asset or self.quote_asset
-		if any_asset:
-			return any_asset
-		raise ValueError("Empty Trade")
-
-	def complete(self) -> Trade:
-		return Trade(**vars(self))
